Move helpers to helper file for data_display app
